+++ static_precompiler/tests/django_settings.py
@@ -15,6 +15,7 @@ STATIC_URL = MEDIA_URL = "/static/"
 
 # noinspection PyUnresolvedReferences
 STATICFILES_DIRS = (
+    os.path.join(os.path.dirname(__file__), 'compilestatic'),
     os.path.join(os.path.dirname(__file__), 'staticfiles_dir'),
     ("prefix", os.path.join(os.path.dirname(__file__), 'staticfiles_dir_with_prefix')),
 )