+++ static_precompiler/tests/test_management.py
@@ -3,27 +3,32 @@ import os
 import pytest
 from django.core import management
 
-from static_precompiler import settings
+import static_precompiler.settings
 from static_precompiler.management.commands import compilestatic
 
 
 def test_get_scanned_dirs():
 
     assert compilestatic.get_scanned_dirs() == sorted([
+        os.path.join(os.path.dirname(__file__), "compilestatic"),
         os.path.join(os.path.dirname(__file__), "staticfiles_dir"),
         os.path.join(os.path.dirname(__file__), "staticfiles_dir_with_prefix"),
-        settings.STATIC_ROOT
+        static_precompiler.settings.STATIC_ROOT,
     ])
 
 
 @pytest.mark.django_db
-def test_compilestatic_command(monkeypatch, tmpdir):
+@pytest.mark.parametrize("verbosity", (0, 1, ))
+def test_compilestatic_command(verbosity, capsys, monkeypatch, tmpdir):
 
+    monkeypatch.setattr("static_precompiler.management.commands.compilestatic.get_scanned_dirs", lambda: (
+        os.path.join(os.path.dirname(__file__), "compilestatic"),
+    ))
     monkeypatch.setattr("static_precompiler.settings.ROOT", tmpdir.strpath)
 
-    management.call_command("compilestatic")
+    management.call_command("compilestatic", verbosity=verbosity)
 
-    output_path = os.path.join(tmpdir.strpath, settings.OUTPUT_DIR)
+    output_path = os.path.join(tmpdir.strpath, static_precompiler.settings.OUTPUT_DIR)
 
     compiled_files = []
     for root, dirs, files in os.walk(output_path):
@@ -33,16 +38,18 @@ def test_compilestatic_command(monkeypatch, tmpdir):
     compiled_files.sort()
 
     assert compiled_files == [
-        "another_test.js",
-        "scripts/test.js",
-        "styles/less/imported.css",
-        "styles/less/test.css",
-        "styles/sass/precision.css",
-        "styles/sass/test.css",
-        "styles/stylus/A.css",
-        "styles/stylus/B/C.css",
-        "styles/stylus/D.css",
-        "styles/stylus/E/F.css",
-        "styles/stylus/E/index.css",
-        "test-compass.css",
+        "coffee/test.js",
+        "less/test.css",
+        "scss/test.css",
     ]
+
+    stdout, _ = capsys.readouterr()
+
+    if verbosity >= 1:
+        assert stdout == (
+            "Compiled 'coffee/test.coffee' to 'COMPILED/coffee/test.js'\n"
+            "Compiled 'less/test.less' to 'COMPILED/less/test.css'\n"
+            "Compiled 'scss/test.scss' to 'COMPILED/scss/test.css'\n"
+        )
+    else:
+        assert stdout == ""